Update JS bundle, body attributes, and add new HTML structure
Updated the <script> tag in the <head> to reference the new
JavaScript bundle `bundle9aa1d002f105eb2fb8b4.js`. Modified
the <body> tag to include an `id` attribute with the value
`body` and a `class` attribute with the value `nis body`.

Added a new HTML structure inside the <body> tag, including
a `div` with the class `nismainborder fullscreen`, containing
nested `div`s with the class `nisborder` and an element with
the `id` of `rotation` and the class `rotation`.