Add "Clear All" button to reset elements and dropdowns
A "Clear All" button was added to the `App` component in `index.js`.
It is styled with the `nisbutton clear-button` class and includes a
trash icon (`<i className="fas fa-trash">`). The button is positioned
next to the "Add New" button with a `marginLeft` of `10px`.

Clicking the button triggers an inline `onClick` function that clears
all elements by calling `setElements([])` and clears all dropdowns by
calling `setDropdowns([])`.